change cpuControllers

+++ MetricsManager/MetricsAgent/DAL/Migrations/FirstMigration.cs
@@ -26,7 +26,7 @@ namespace MetricsAgent.DAL.Migrations
                 .WithColumn("time").AsInt64();
             Create.Table("hddmetrics")
                 .WithColumn("id").AsInt64().PrimaryKey().Identity()
-                .WithColumn("value").AsInt32()
+                .WithColumn("value").AsInt64()
                 .WithColumn("time").AsInt64();
             Create.Table("networkmetrics")
                 .WithColumn("id").AsInt64().PrimaryKey().Identity()
